Calculations class added
