+++ .idea/.idea.cycling-map/.idea/workspace.xml
@@ -7,13 +7,10 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="86726afe-2fd4-4c55-84e0-de4b041a8532" name="Changes" comment="">
-      <change afterPath="$PROJECT_DIR$/cycling-map/IParser.cs" afterDir="false" />
+    <list default="true" id="86726afe-2fd4-4c55-84e0-de4b041a8532" name="Changes" comment="new interface Parser">
+      <change afterPath="$PROJECT_DIR$/cycling-map/MathCalculations.cs" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/.idea.cycling-map/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/.idea.cycling-map/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/cycling-map/CalculateRoute.cs" beforeDir="false" afterPath="$PROJECT_DIR$/cycling-map/CalculateRoute.cs" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/cycling-map/GeocodeParser.cs" beforeDir="false" afterPath="$PROJECT_DIR$/cycling-map/GeocodeParser.cs" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/cycling-map/MainWindow.xaml.cs" beforeDir="false" afterPath="$PROJECT_DIR$/cycling-map/MainWindow.xaml.cs" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/cycling-map/RouteParser.cs" beforeDir="false" afterPath="$PROJECT_DIR$/cycling-map/RouteParser.cs" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -49,6 +46,7 @@
     <setting file="file://$USER_HOME$/AppData/Local/Symbols/src/dotnet/runtime/2d7eea252964e69be94cb9c847b371b23e4dd470/src/libraries/System.Private.CoreLib/src/System/Threading/ExecutionContext.cs" root0="SKIP_HIGHLIGHTING" />
     <setting file="file://$APPLICATION_CONFIG_DIR$/resharper-host/DecompilerCache/decompiler/9e4675b272dd4b9ab305a68d4d11279ff618a0/84/c605dcf3/MessageBox.cs" root0="SKIP_HIGHLIGHTING" />
     <setting file="file://$APPLICATION_CONFIG_DIR$/resharper-host/SourcesCache/4bda76b5cc453e1edf5d5c754c4a8215edbd3d3e4f80706dcf4f52a4f68979/EventRoute.cs" root0="SKIP_HIGHLIGHTING" />
+    <setting file="mock://C:/Users/obbte/RiderProjects/cycling-map/cycling-map/CalculateRoute.cs" root0="SKIP_HIGHLIGHTING" />
     <setting file="mock://C:/Users/obbte/RiderProjects/cycling-map/cycling-map/GeocodeParser.cs" root0="SKIP_HIGHLIGHTING" />
     <setting file="mock://C:/Users/obbte/RiderProjects/cycling-map/cycling-map/GeocodeParser.cs" root0="SKIP_HIGHLIGHTING" />
   </component>
@@ -68,7 +66,7 @@
   "keyToString": {
     ".NET Project.cycling-map.executor": "Run",
     "RunOnceActivity.ShowReadmeOnStart": "true",
-    "git-widget-placeholder": "illia-test-obidkhon",
+    "git-widget-placeholder": "obidkhon-implement-calculate-class",
     "node.js.detected.package.eslint": "true",
     "node.js.detected.package.tslint": "true",
     "node.js.selected.package.eslint": "(autodetect)",
@@ -117,7 +115,7 @@
       <workItem from="1718584016905" duration="1721000" />
       <workItem from="1718626912459" duration="2299000" />
       <workItem from="1718881289643" duration="1666000" />
-      <workItem from="1719408931334" duration="4925000" />
+      <workItem from="1719408931334" duration="7668000" />
     </task>
     <task id="LOCAL-00001" summary="first commit">
       <option name="closed" value="true" />
@@ -151,7 +149,15 @@
       <option name="project" value="LOCAL" />
       <updated>1718568357572</updated>
     </task>
-    <option name="localTasksCounter" value="5" />
+    <task id="LOCAL-00005" summary="new interface Parser">
+      <option name="closed" value="true" />
+      <created>1719414013768</created>
+      <option name="number" value="00005" />
+      <option name="presentableId" value="LOCAL-00005" />
+      <option name="project" value="LOCAL" />
+      <updated>1719414013768</updated>
+    </task>
+    <option name="localTasksCounter" value="6" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -188,7 +194,8 @@
     <MESSAGE value="test two integrate calculate route with map rendering" />
     <MESSAGE value="working route calculation&#10;working map and route rendering" />
     <MESSAGE value="start and end points bug fix" />
-    <option name="LAST_COMMIT_MESSAGE" value="start and end points bug fix" />
+    <MESSAGE value="new interface Parser" />
+    <option name="LAST_COMMIT_MESSAGE" value="new interface Parser" />
   </component>
   <component name="XDebuggerManager">
     <watches-manager>
